+++ TerrasRebalance/Plugin.cs
@@ -6,6 +6,7 @@ using System.IO;
 using System.Collections.Generic;
 using UnityEngine.Tilemaps;
 using System.Dynamic;
+using BepInEx.Configuration;
 
 namespace TerrasRebalance;
 
@@ -19,32 +20,58 @@ public class Plugin : BaseUnityPlugin
     {
         Instance ??= this;
 
+        Configuration.Bind(base.Config);
+
         Logger.LogInfo($"Plugin {PluginInfo.PLUGIN_GUID} is loaded!");
 
-        Logger.LogInfo("Applying Bees Patch");
-        harmony.PatchAll(typeof(CircuitBeesPatch));
-        harmony.PatchAll(typeof(ButlerBeesPatch));
-        Logger.LogInfo("Bees Patched");
+        if (Configuration.changeBeeBehavior.Value)
+        {
+            Logger.LogInfo("Applying Bees Patch");
+            harmony.PatchAll(typeof(CircuitBeesPatch));
+            harmony.PatchAll(typeof(ButlerBeesPatch));
+            Logger.LogInfo("Bees Patched");
+        }
+        else Logger.LogInfo("Skipping Bees Patch");
 
-        Logger.LogInfo("Applying Butler Patch");
-        harmony.PatchAll(typeof(ButlerPatch));
-        Logger.LogInfo("Butler Patched");
+        if (Configuration.changeButlerHealth.Value)
+        {
+            Logger.LogInfo("Applying Butler Patch");
+            harmony.PatchAll(typeof(ButlerPatch));
+            Logger.LogInfo("Butler Patched");
+        }
+        else Logger.LogInfo("Skipping Butler Patch");
 
-        Logger.LogInfo("Applying Jester Patch");
-        harmony.PatchAll(typeof(JesterPatch));
-        Logger.LogInfo("Jester Patched");
+        if (Configuration.changeJesterTimer.Value)
+        {
+            Logger.LogInfo("Applying Jester Patch");
+            harmony.PatchAll(typeof(JesterPatch));
+            Logger.LogInfo("Jester Patched");
+        }
+        else Logger.LogInfo("Skipping Jester Patch");
 
-        Logger.LogInfo("Applying Nutcracker Patch");
-        harmony.PatchAll(typeof(NutcrackerPatch));
-        Logger.LogInfo("Nutcracker Patched");
+        if (Configuration.changeNutcrackerBehavior.Value)
+        {
+            Logger.LogInfo("Applying Nutcracker Patch");
+            harmony.PatchAll(typeof(NutcrackerPatch));
+            Logger.LogInfo("Nutcracker Patched");
+        }
+        else Logger.LogInfo("Skipping Nutcracker Patch");
 
-        Logger.LogInfo("Applying Shovel Patch");
-        harmony.PatchAll(typeof(ShovelPatch));
-        Logger.LogInfo("Shovel Patched");
+        if (Configuration.changeShovelWeight.Value || Configuration.changeStopSignWeight.Value || Configuration.changeJetpackPrice.Value)
+        {
+            Logger.LogInfo("Applying Items Patch");
+            harmony.PatchAll(typeof(ItemPatch));
+            Logger.LogInfo("Items Patched");
+        }
+        else Logger.LogInfo("Skipping Items Patch");
 
-        Logger.LogInfo("Applying Stormy Patch");
-        harmony.PatchAll(typeof(StormyPatch));
-        Logger.LogInfo("Stormy Patched");
+        if (Configuration.changeStormyTimers.Value)
+        {
+            Logger.LogInfo("Applying Stormy Patch");
+            harmony.PatchAll(typeof(StormyPatch));
+            Logger.LogInfo("Stormy Patched");
+        }
+        else Logger.LogInfo("Skipping Stormy Patch");
     }
 
     public static void Log(string msg) => Instance.Logger.LogInfo(msg);