+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+# Version 1.1.0
+- Added a toggle for each of the changes.
+- Added a change for Stop Signs to be 21 lbs.
+- Added a change for Jetpacks to be $700.
+- Stormy rebalance has been nerfed. Strikes are now only 25% more spaced out instead of 50%.
+- Removed Jesters slamming open doors.
+
 # Version 1.0.1
 - Changed Butler's HP to be only 4 on MP.